Fix problem in LB using check_cache thread (cacheex build). See forum: http://www.streamboard.tv/wbb2/thread.php?postid=491986#post491986 Add cw counter info in log, when getting cw from cache and cw counter>1.
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@9144 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -860,7 +860,7 @@ typedef struct ecm_request_t
 	uint8_t         cacheex_hitcache;           // =1 if wait_time due hitcache
 	void            *cw_cache;					//pointer to cw stored in cache
 #endif
-	uint8_t         cw_count;
+	uint32_t        cw_count;
 	uint8_t         from_csp;                   // =1 if er from csp cache
 	uint8_t         from_cacheex;               // =1 if er from cacheex client pushing cache
 	uint8_t         from_cacheex1_client;       // =1 if er from cacheex-1 client
